Module espresso_esqs::full_node

Full node query service

An instantiation of a Tide Disco service containing all possible query service modules, assuming all the relevant data is available locally. This service assumes it is running in the same process as an Espresso/HotShot instance. It provides the following API modules:

availability

catchup

metastate

status

validator

Structs

EsQS

Options

UpdateQueryDataSourceTypesBinder 🔒

Enums

Command

Functions

io_error 🔒